rename file to git helpers
