+++ protostar/git/__init__.py
@@ -1,4 +1,4 @@
-from .git import get_git_version
+from .git_helpers import get_git_version
 from .git_repository import GitRepository, Submodule
 from .git_exceptions import (
     GitNotFoundException,
